backend/requirements.txt
# Pinned for Python 3.12 (Docker). Local dev on Python 3.9 may show starlette CVEs
# that are fixed in starlette>=0.47.2 — those are resolved automatically on Python 3.12.
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.9.1
argon2-cffi==23.1.0
slowapi==0.1.9
sqlalchemy==2.0.40
greenlet==3.1.1
aiosqlite==0.20.0
httpx==0.28.1
aiomqtt==2.3.0
apscheduler==3.10.4
pytest==8.3.5
pytest-asyncio==0.24.0
